+++ scripts/check.sh
@@ -1,5 +1,6 @@
 set -e
 set -x
+./gradlew generator:jvmTest
 ./gradlew viewmodel:jvmTest
 ./gradlew model:jvmTest
 ./gradlew viewlegacy:compileDebugAndroidTestSources